Disabled No Magic Numbers.....
My gosh. They're so frustrating everywhere.

BUT - Gave you `--annoy-mode` to add on top of your standards if you
want it.

Added in test-override, which over the top of your normal rules, lets
you have unlimited nesting and unused expressions without issue.
Probably not a complete set, but a start.